@@ -8,7 +8,7 @@ import sys
 
 # openjdk15 is only for bootstrapping openjdk
 releases = ("openjdk8", "openjdk11", "openjdk13", "openjdk14", "openjdk15", "openjdk16", "openjdk17")
-oses = ("mac", "linux")
+oses = ("mac", "linux", "alpine_linux")
 types = ("jre", "jdk")
 impls = ("hotspot", "openj9")
 
