@@ -6,8 +6,8 @@ import re
 import requests
 import sys
 
-feature_versions = (8, 11, 16, 17)
-oses = ("mac", "linux")
+feature_versions = (8, 11, 16, 17, 18, 19)
+oses = ("mac", "linux", "alpine-linux")
 types = ("jre", "jdk")
 impls = ("hotspot")
 
@@ -15,6 +15,7 @@ arch_to_nixos = {
     "x64": ("x86_64",),
     "aarch64": ("aarch64",),
     "arm": ("armv6l", "armv7l"),
+    "ppc64le": ("powerpc64le",),
 }
 
 def generate_sources(assets, feature_version, out):
