@@ -1,4 +1,4 @@
-const path = require('path')
+const path = require("path");
 
 // String -> String
 
@@ -10,20 +10,19 @@ const path = require('path')
 // - https://codeload.github.com/Gargron/emoji-mart/tar.gz/934f314fd8322276765066e8a2a6be5bac61b1cf
 
 function urlToName(url) {
-
   // Yarn generates `codeload.github.com` tarball URLs, where the final
   // path component (file name) is the git hash. See #111.
   // See also https://github.com/yarnpkg/yarn/blob/989a7406/src/resolvers/exotics/github-resolver.js#L24-L26
   let isCodeloadGitTarballUrl =
-    url.startsWith('https://codeload.github.com/') && url.includes('/tar.gz/')
+    url.startsWith("https://codeload.github.com/") && url.includes("/tar.gz/");
 
-  if (url.startsWith('git+') || isCodeloadGitTarballUrl) {
-    return path.basename(url)
+  if (url.startsWith("git+") || isCodeloadGitTarballUrl) {
+    return path.basename(url);
   }
 
   return url
-    .replace(/https:\/\/(.)*(.com)\//g, '') // prevents having long directory names
-    .replace(/[@/%:-]/g, '_') // replace @ and : and - and % characters with underscore
+    .replace(/https:\/\/(.)*(.com)\//g, "") // prevents having long directory names
+    .replace(/[@/%:-]/g, "_"); // replace @ and : and - and % characters with underscore
 }
 
-module.exports = urlToName
+module.exports = urlToName;
