@@ -25,18 +25,18 @@ def nix_prefetch_url(url, algo='sha256'):
     """Prefetches the content of the given URL."""
     print(f'nix-prefetch-url {url}')
     out = subprocess.check_output(['nix-prefetch-url', '--type', algo, url])
-    return out.decode('utf-8').rstrip()
+    return out.rstrip()
 
 
 current_version = subprocess.check_output([
     'nix', 'eval', '--raw',
     '-f', dirname(abspath(__file__)) + '/../../../..',
     'dictdDBs.wiktionary.version',
-]).decode('utf-8')
+])
 
 parser = WiktionaryLatestVersionParser(current_version)
 
 with urlopen('https://dumps.wikimedia.org/enwiktionary/') as resp:
-    parser.feed(resp.read().decode('utf-8'))
+    parser.feed(resp.read())
 
 print(parser.latest_version)
